requirements.txt
torch
torchvision
streamlit
streamlit-webrtc
Pillow
matplotlib
streamlit-webrtc
